@@ -1,13 +1,19 @@
 {
-  "name": "nmdc-server",
-  "version": "0.1.0",
+  "name": "web",
   "private": true,
+  "type": "module",
+  "version": "0.0.0",
   "scripts": {
-    "serve": "rimraf -rf ./node_modules/.cache/vue-loader && vue-cli-service serve --host 127.0.0.1 --port 8081 src/main.ts",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --fix"
   },
   "dependencies": {
+    "@fontsource/roboto": "5.2.7",
+    "@mdi/font": "7.4.47",
     "@riophae/vue-treeselect": "^0.4.0",
     "@sentry/vue": "^6.16.1",
     "axios": "^0.28.0",
@@ -37,68 +43,32 @@
     "popper.js": "1.16.1",
     "protobufjs": "^6.11.3",
     "serialize-javascript": "^6.0.0",
-    "vue": "^2.7.0",
+    "vue": "^3.5.21",
     "vue-async-computed": "^3.9.0",
     "vue-google-charts": "^0.3.3",
     "vue-gtag": "^1.16.1",
-    "vue-router": "^3.6.0",
+    "vue-router": "^4.5.1",
     "vue2-leaflet": "^2.7.1",
     "vue2-leaflet-markercluster": "^3.1.0",
-    "vuetify": "^2.7.0",
+    "vuetify": "^3.10.1",
     "xlsx": "^0.17.5"
   },
   "devDependencies": {
-    "@types/axios": "^0.14.0",
-    "@types/bonjour": "^3.5.13",
-    "@types/connect": "^3.4.38",
-    "@types/connect-history-api-fallback": "^1.5.4",
-    "@types/d3": "^6.7.6",
-    "@types/d3-array": "^2.12.1",
-    "@types/d3-brush": "^2.1.5",
-    "@types/d3-scale": "^3.3.4",
-    "@types/d3-selection": "^2.0.4",
-    "@types/d3-time-format": "^3.0.1",
-    "@types/estree": "^1.0.8",
-    "@types/geojson": "^7946.0.16",
-    "@types/leaflet": "^1.7.6",
-    "@types/linkify-it": "^3.0.2",
-    "@types/lodash": "^4.14.178",
-    "@typescript-eslint/eslint-plugin": "^5.6.0",
-    "@typescript-eslint/parser": "^5.6.0",
-    "@vue/cli-plugin-babel": "^5.0.0",
-    "@vue/cli-plugin-eslint": "^5.0.0",
-    "@vue/cli-plugin-typescript": "^5.0.0",
-    "@vue/cli-service": "^5.0.9",
-    "@vue/eslint-config-airbnb": "^6.0.0",
-    "@vue/eslint-config-typescript": "^9.1.0",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^10.0.0",
-    "copy-webpack-plugin": "^6.4.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.25.3",
-    "eslint-plugin-vue": "^8.2.0",
-    "eslint-plugin-vuejs-accessibility": "^1.1.0",
-    "process": "^0.11.10",
-    "querystring-es3": "^0.2.1",
-    "raw-loader": "^4.0.2",
-    "sass": "~1.32.0",
-    "sass-loader": "^10.1.1",
-    "thread-loader": "3.0.4",
-    "typescript": "^4.5.3",
-    "vue-cli-plugin-vuetify": "^2.4.5",
-    "vue-cli-plugin-yaml": "^1.0.2",
-    "vue-template-compiler": "^2.7.0",
-    "vuetify-loader": "^1.7.3",
-    "webpack": "^5.101.3"
+    "@tsconfig/node22": "^22.0.0",
+    "@types/node": "^22.9.0",
+    "@vitejs/plugin-vue": "^6.0.1",
+    "@vue/tsconfig": "^0.8.1",
+    "eslint": "^9.35.0",
+    "eslint-config-vuetify": "^4.2.0",
+    "npm-run-all2": "^8.0.4",
+    "sass-embedded": "^1.92.1",
+    "typescript": "~5.9.2",
+    "unplugin-fonts": "^1.4.0",
+    "unplugin-vue-components": "^29.0.0",
+    "unplugin-vue-router": "^0.15.0",
+    "vite": "^7.1.5",
+    "vite-plugin-vuetify": "^2.1.2",
+    "vue-tsc": "^3.0.7"
   },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ],
-  "resolutions": {
-    "node-forge": "^1.3.0",
-    "serialize-javascript": "^3.1",
-    "http-proxy": "1.18.1"
-  },
-  "packageManager": "yarn@1.22.19+sha1.4ba7fc5c6e704fce2066ecbfb0b0d8976fe62447"
+  "packageManager": "yarn@1.22.22+sha256.c17d3797fb9a9115bf375e31bfd30058cac6bc9c3b8807a3d8cb2094794b51ca"
 }
